backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
fastapi==0.115.12
gcloud==0.18.3
googleapis-common-protos==1.69.2
greenlet==3.1.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
jwcrypto==1.5.6
oauth2client==4.1.3
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
Pyrebase4==4.8.0
python-jwt==4.1.0
python-multipart==0.0.20
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==1.26.20
uvicorn==0.34.0
